LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audiobuffer-interface/idl-test-expected.txt

CONSOLE MESSAGE: line 331: callback not yet supported
CONSOLE MESSAGE: line 331: callback not yet supported
dictionary AudioBufferOptions {
             unsigned long numberOfChannels = 1;
    required unsigned long length;
    required float         sampleRate;
};

[Constructor(AudioBufferOptions options)]
interface AudioBuffer {
    readonly        attribute float         sampleRate;
    readonly        attribute unsigned long length;
    readonly        attribute double        duration;
    readonly        attribute unsigned long numberOfChannels;
    Float32Array getChannelData (unsigned long channel);
    void         copyFromChannel (Float32Array destination, unsigned long channelNumber, optional unsigned long startInChannel = 0
              );
    void         copyToChannel (Float32Array source, unsigned long channelNumber, optional unsigned long startInChannel = 0
              );
};

PASS AudioBuffer interface: existence and properties of interface object 
FAIL AudioBuffer interface object length assert_equals: wrong value for AudioBuffer.length expected 1 but got 0
PASS AudioBuffer interface object name 
PASS AudioBuffer interface: existence and properties of interface prototype object 
PASS AudioBuffer interface: existence and properties of interface prototype object's "constructor" property 
PASS AudioBuffer interface: existence and properties of interface prototype object's @@unscopables property 
PASS AudioBuffer interface: attribute sampleRate 
PASS Unscopable handled correctly for sampleRate property on AudioBuffer 
PASS AudioBuffer interface: attribute length 
PASS Unscopable handled correctly for length property on AudioBuffer 
PASS AudioBuffer interface: attribute duration 
PASS Unscopable handled correctly for duration property on AudioBuffer 
PASS AudioBuffer interface: attribute numberOfChannels 
PASS Unscopable handled correctly for numberOfChannels property on AudioBuffer 
PASS AudioBuffer interface: operation getChannelData(unsigned long) 
PASS Unscopable handled correctly for getChannelData(unsigned long) on AudioBuffer 
FAIL AudioBuffer interface: operation copyFromChannel(Float32Array, unsigned long, unsigned long) assert_own_property: interface prototype object missing non-static operation expected property "copyFromChannel" missing
PASS Unscopable handled correctly for copyFromChannel(Float32Array, unsigned long, unsigned long) on AudioBuffer 
FAIL AudioBuffer interface: operation copyToChannel(Float32Array, unsigned long, unsigned long) assert_own_property: interface prototype object missing non-static operation expected property "copyToChannel" missing
PASS Unscopable handled correctly for copyToChannel(Float32Array, unsigned long, unsigned long) on AudioBuffer 
PASS AudioBuffer must be primary interface of audio_buffer 
PASS Stringification of audio_buffer 
PASS AudioBuffer interface: audio_buffer must inherit property "sampleRate" with the proper type 
PASS AudioBuffer interface: audio_buffer must inherit property "length" with the proper type 
PASS AudioBuffer interface: audio_buffer must inherit property "duration" with the proper type 
PASS AudioBuffer interface: audio_buffer must inherit property "numberOfChannels" with the proper type 
PASS AudioBuffer interface: audio_buffer must inherit property "getChannelData(unsigned long)" with the proper type 
PASS AudioBuffer interface: calling getChannelData(unsigned long) on audio_buffer with too few arguments must throw TypeError 
FAIL AudioBuffer interface: audio_buffer must inherit property "copyFromChannel(Float32Array, unsigned long, unsigned long)" with the proper type assert_inherits: property "copyFromChannel" not found in prototype chain
FAIL AudioBuffer interface: calling copyFromChannel(Float32Array, unsigned long, unsigned long) on audio_buffer with too few arguments must throw TypeError assert_inherits: property "copyFromChannel" not found in prototype chain
FAIL AudioBuffer interface: audio_buffer must inherit property "copyToChannel(Float32Array, unsigned long, unsigned long)" with the proper type assert_inherits: property "copyToChannel" not found in prototype chain
FAIL AudioBuffer interface: calling copyToChannel(Float32Array, unsigned long, unsigned long) on audio_buffer with too few arguments must throw TypeError assert_inherits: property "copyToChannel" not found in prototype chain
